+++ Functions/cmdlet.ps1
@@ -1,5 +1,7 @@
 Set-Alias gbpm Get-BoundParams
 Set-Alias gcp Get-CommonParams
+Set-Alias '>>' ConvertTo-ParamObject
+Set-Alias icms Invoke-CommandSafely
 
 Function Get-BoundParams
 {
@@ -293,4 +295,237 @@ Test-ValidFileName
 
         return [scriptblock]::Create($code)
     }
+}
+Function ConvertTo-ParamObject
+{
+    [CmdletBinding()]
+    param
+    (
+        [parameter(Position          = 1,
+                   Mandatory         = $true,
+                   ValueFromPipeline = $true)]
+        $InputObject
+    )
+    process
+    {
+        if
+        (
+            # the usual type of splat parameters
+            $InputObject -is [hashtable] -or
+
+            # the type of PSBoundParameters
+            ([string]$InputObject.GetType()) -eq 'System.Collections.Generic.Dictionary`2[System.String,System.Object]]'
+        )
+        {
+            return New-Object psobject -Property $InputObject
+        }
+
+        return $InputObject
+    }
+}
+
+if ( $PSVersionTable.PSVersion.Major -ge 4)
+{
+function Get-Parameters
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Position                        = 1,
+                   Mandatory                       = $true,
+                   ValueFromPipeline               = $true,
+                   ValueFromPipelineByPropertyName = $true)]
+        [string]
+        $CmdletName,
+
+        [Parameter(Position                        = 2,
+                   ValueFromPipelineByPropertyName = $true)]
+        [string]
+        $ParameterSetName,
+
+        [Parameter(Position                        = 3,
+                   ValueFromPipelineByPropertyName = $true)]
+        [ValidateSet('Required','All')]
+        [string]
+        $Mode='All'
+    )
+    process
+    {
+        $bp = &(gbpm)
+
+        $cmd = Get-Command $CmdletName -ErrorAction Stop
+
+        if
+        (
+             $bp.Keys -notcontains 'ParameterSetName' -and
+            $cmd.ParameterSets.Count -gt 1 -and
+            -not $cmd.DefaultParameterSet
+        )
+        {
+            throw New-Object System.ArgumentException(
+                "Cmdlet $CmdletName has more than one parameterset and no default. You must provide ParameterSetName.",
+                'ParameterSetName'
+            )
+        }
+
+        if
+        (
+            $bp.Keys -contains 'ParameterSetName' -and
+            ($cmd.ParameterSets | % {$_.Name}) -notcontains $ParameterSetName
+        )
+        {
+            throw New-Object System.ArgumentException(
+                "Cmdlet $CmdletName does not have ParameterSetName $ParameterSetName.",
+                'ParameterSetName'
+            )
+        }
+
+        if ( -not ($cmd.Parameters.Keys | ? { (Get-CommonParameterNames) -notcontains $_}))
+        {
+            return
+        }
+
+        if
+        (
+            $bp.Keys -notcontains 'ParameterSetName' -and
+            $cmd.DefaultParameterSet
+        )
+        {
+            $psn = $cmd.DefaultParameterSet
+        }
+
+        if
+        (
+            $bp.Keys -notcontains 'ParameterSetName' -and
+            -not $cmd.DefaultParameterSet
+        )
+        {
+            $psn = $cmd.ParameterSets[0].Name
+        }
+
+        if
+        (
+            $bp.Keys -contains 'ParameterSetName'
+        )
+        {
+            $psn = $ParameterSetName
+        }
+
+        $parameterSet = $cmd.ParameterSets | ? { $_.Name -eq $psn }
+        $parameterSetParameterNames = $parameterSet.Parameters |
+            ? { (Get-CommonParameterNames) -notcontains $_.Name } |
+            % {$_.Name}
+
+        $help = Get-Help $CmdletName
+
+        return $help.parameters.parameter |
+            ? {
+                (
+                    $_.Required -eq $true -or
+                    $Mode -eq 'All'
+                ) -and
+                $parameterSetParameterNames -contains $_.Name
+            } |
+            % {$_.Name}
+    }
+}
+function Test-ValidParams
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Position                        = 1,
+                   Mandatory                       = $true,
+                   ValueFromPipelineByPropertyName = $true)]
+        [string]
+        $CmdletName,
+
+        [Parameter(Position                        = 2,
+                   Mandatory                       = $true,
+                   ValueFromPipelineByPropertyName = $true)]
+        [hashtable]
+        $SplatParams,
+
+        [Parameter(Position                        = 3,
+                   ValueFromPipelineByPropertyName = $true)]
+        [string]
+        $ParameterSetName,
+
+        [Parameter(Position                        = 4,
+                   ValueFromPipelineByPropertyName = $true)]
+        [string]
+        [ValidateSet('Error','Verbose','Throw')]
+        [Alias('fa')]
+        $FailAction='Verbose'
+    )
+    process
+    {
+        foreach ( $requiredParam in (&(gbpm) | >> | Get-Parameters -Mode Required) )
+        {
+            if ( $SplatParams.Keys -notcontains $requiredParam )
+            {
+                $message = "Required parameter $requiredParam not in SplatParams for Cmdlet $CmdletName"
+                $param = $requiredParam
+            }
+        }
+
+        $allParams = (&(gbpm) | >> | Get-Parameters -Mode All)
+
+        foreach ( $splatParam in $SplatParams.Keys )
+        {
+            if (  $allParams -notcontains $splatParam )
+            {
+                $message = "SplatParam $splatParam provided but not a parameter of Cmdlet $CmdletName"
+                $param = $splatParam
+            }
+        }
+
+        if ($message)
+        {
+            Switch ($FailAction) {
+                'Throw' {
+                    throw New-Object System.ArgumentException(
+                        $message,
+                        $param
+                    )
+                }
+                'Error'   { Write-Error   $message}
+                'Verbose' { Write-Verbose $message }
+            }
+            return $false
+        }
+
+        return $true
+    }
+}
+function Invoke-CommandSafely
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter(Position                        = 1,
+                   Mandatory                       = $true,
+                   ValueFromPipelineByPropertyName = $true)]
+        [string]
+        $CmdletName,
+
+        [Parameter(Position                        = 2,
+                   Mandatory                       = $true,
+                   ValueFromPipelineByPropertyName = $true)]
+        [hashtable]
+        $SplatParams,
+
+        [Parameter(Position                        = 3,
+                   ValueFromPipelineByPropertyName = $true)]
+        [string]
+        $ParameterSetName
+    )
+    process
+    {
+        $bp = &(gbpm)
+        Test-ValidParams @bp -fa Throw | Out-Null
+
+        & $CmdletName @SplatParams
+    }
+}
 }